refactor(lecture): 총원 필드 비정규화

+++ src/main/java/org/hyejoon/cuvcourse/domain/lecture/lectureView/dto/LectureViewResponse.java
@@ -1,5 +1,5 @@
 package org.hyejoon.cuvcourse.domain.lecture.lectureView.dto;
 
 public record LectureViewResponse(Long id, String lectureTitle, String professorName, int credits,
-                                  long totalStudents, int capacity) {
+                                  int totalStudents, int capacity) {
 }
